% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_account.R
\name{pardot_account}
\alias{pardot_account}
\title{Retrieve Pardot Account}
\usage{
pardot_account()
}
\value{
A tbl_df data frame.
}
\description{
Make a call to the Pardot API and return the account data.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_account()}
}
