% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_campaigns.R
\name{pardot_campaigns}
\alias{pardot_campaigns}
\title{Make a call to the Pardot API and return a dataframe listing all campaigns}
\usage{
pardot_campaigns()
}
\description{
Make a call to the Pardot API and return a dataframe listing all campaigns
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
pardot_campaigns()}
}

