% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_client.R
\name{pardot_client}
\alias{pardot_client}
\title{Make a call to the Pardot API and return XML}
\usage{
pardot_client(object, operator, identifier_field = NULL, identifier = NULL)
}
\arguments{
\item{object}{A string containing a Pardot Object}

\item{operator}{A string containing a Pardot Operator}

\item{identifier_field}{A string with an optional identifier field. Can be null}

\item{identifier}{A string with an optional identifier that can be null if identifier_field is null}
}
\description{
Make a call to the Pardot API and return XML
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
pardot_client("campaigns", "query")}
}

