% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_prospects.R
\name{pardot_prospects}
\alias{pardot_prospects}
\title{Make a call to the Pardot API and return a dataframe listing all prospects}
\usage{
pardot_prospects()
}
\description{
Make a call to the Pardot API and return a dataframe listing all prospects
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
pardot_prospects()}
}

