% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_visitors.R
\name{pardot_visitors}
\alias{pardot_visitors}
\title{Retrieve Pardot Visitors}
\usage{
pardot_visitors(...)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, name, update_before, 
updated_after, only_identified, prospect_ids.}
}
\value{
A tbl_df data frame.
}
\description{
Make a call to the Pardot API and returns the visitors matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_visitors()
df <- pardot_visitors(created_after = 'yesterday', only_identified = 'true') }
}
