% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_visits.R
\name{pardot_visits}
\alias{pardot_visits}
\title{Retrieve Pardot Visits}
\usage{
pardot_visits(...)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are ids, visitor_ids, prospect_ids.}
}
\value{
A tbl_df data frame. Note: The field visitor_page_view contains nested data frames (to be fixed).
}
\description{
Make a call to the Pardot API and returns the visits matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_visits()
df <- pardot_visits(created_after = 'yesterday', prospect_id = 123) }
}
