% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_curl_options.R
\name{set_curl_options}
\alias{set_curl_options}
\title{Set curl options for Pardot API calls
See httr::config() and curl::handle_setopt() for options.}
\usage{
set_curl_options(curl_options = NULL)
}
\arguments{
\item{curl_options}{To set curl options directly. Usage: curl_options = httr:config(...)}
}
\description{
Set curl options for Pardot API calls
See httr::config() and curl::handle_setopt() for options.
}
\examples{
\dontrun{
set_curl_options(httr::config(sslversion = 6)) # Connect using SSL/TLS v1.2
set_credentials("your-username", "your-password", "your-user-key")}
}
