% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_oauth_credentials.R
\name{set_oauth_credentials}
\alias{set_oauth_credentials}
\title{Set the Pardot user credentials and authenticate using Salesforce OAuth}
\usage{
set_oauth_credentials(pardot_username, pardot_password,
  pardot_businessunit_id, client_id, client_secret)
}
\arguments{
\item{pardot_username}{A string containing your Pardot UserName}

\item{pardot_password}{A string containing your Pardot Password}

\item{pardot_businessunit_id}{A string containing your Pardot Business Unit Id}

\item{client_id}{A string containing your Salesforce OAuth client id}

\item{client_secret}{A string containing your Salesforce OAuth client_secret}
}
\value{
Authentication result (a token or an error message)
}
\description{
Authenticate a user using Salesforce OAuth Access Token method.
}
\examples{
\dontrun{
set_oauth_credentials("username", "password", "businessunit-id", "client-id", "client-secret")}
}
